social_media_api/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
Pillow>=9.0.0
